fix: docs now work (#405)
* finally

* path, and copy

* oops haha

+++ src/app/system/sync/fileCache.ts
@@ -102,7 +102,7 @@ export const downloadFileToCache = async ({
     }
 
     const urlMap = readAtom(urlMapAtom)
-    if (urlMap[url] !== undefined) {
+    if (urlMap[url] !== undefined && urlMap[url] !== filename) {
       try {
         await unlink(getFilePath({ filename: urlMap[url], type }))
         // eslint-disable-next-line no-empty
